Fix timer races (#1)
* Non-blocking write to the timer channel

In case the user calls Reset without draining the channel.

* Fix tick/reset/stop race

We release the lock before ticking which means the timer could have been
reset. This can lead to the following issue:

1. Something triggers a timer, calling `Tick`.
2. Concurrently, `Stop` is called. `Stop` returns false because the
timer was running and we succeed in stopping the timer.
3. `Tick` takes the lock and triggers the timer anyways.

The fix is to check if we're stopped inside the timer. Maybe we should
avoid dropping the lock in the first place but I want to focus on the
direct fix for now.